requirements.txt
# Core
streamlit==1.35.0
python-dotenv==1.0.1
pandas==2.2.2

# LLM (Groq via OpenAI-compatible client)
openai==1.30.1

# Agents
langgraph==0.2.16
langchain-core==0.2.7

# Embeddings (local, pas de clé API)
sentence-transformers==3.0.1

# Vector DB
qdrant-client==1.9.1

# Spotify
spotipy==2.23.0

# Ingestion
dlt==0.4.12

# Evaluation
numpy==1.26.4

# Dev
pytest==8.2.2
